Split interfaces to allow for better abstraction
