+++ corehq/apps/hqcase/tests/test_case_assigment.py
@@ -9,7 +9,8 @@ from corehq.apps.hqcase.exceptions import CaseAssignmentError
 from corehq.apps.hqcase.utils import assign_case
 from corehq.apps.users.models import CommCareUser
 from corehq.apps.users.util import format_username
-from corehq.form_processor.interfaces import FormProcessorInterface
+from corehq.form_processor.interfaces.case import CaseInterface
+from corehq.form_processor.interfaces.processor import FormProcessorInterface
 
 
 class CaseAssignmentTest(TestCase):
@@ -102,7 +103,7 @@ class CaseAssignmentTest(TestCase):
         # shouldn't fail
         case = self._new_case()
         case_with_bad_ref = self._new_case(index={'parent': ('person', case._id)})
-        FormProcessorInterface.soft_delete_case(case._id)
+        CaseInterface.soft_delete(case._id)
         # this call previously failed
         res = assign_case(case_with_bad_ref.id, self.primary_user._id,
                           include_subcases=True, include_parent_cases=True)
@@ -136,13 +137,13 @@ class CaseAssignmentTest(TestCase):
     def _check_state(self, new_owner_id, expected_changed, update=None):
         expected_ids = set(c._id for c in expected_changed)
         for case in expected_changed:
-            expected = FormProcessorInterface.get_case(case._id)
+            expected = CaseInterface.get_case(case._id)
             self.assertEqual(new_owner_id, expected.owner_id)
             if update:
                 for prop, value in update.items():
                     self.assertEqual(getattr(expected, prop), value)
         for case in (c for c in self.all if c._id not in expected_ids):
-            remaining = FormProcessorInterface.get_case(case._id)
+            remaining = CaseInterface.get_case(case._id)
             self.assertEqual(self.original_owner._id, remaining.owner_id)
 
     def _new_case(self, index=None):
