+++ corehq/apps/userreports/tests/test_view.py
@@ -12,7 +12,7 @@ from casexml.apps.case.util import post_case_blocks
 from casexml.apps.case.xml import V2
 from corehq.apps.userreports.reports.view import ConfigurableReport
 from corehq.db import Session
-from corehq.form_processor.interfaces import FormProcessorInterface
+from corehq.form_processor.interfaces.processor import FormProcessorInterface
 
 
 class ConfigurableReportViewTest(TestCase):
