+++ corehq/ex-submodules/casexml/apps/case/mock.py
@@ -3,7 +3,8 @@ import copy
 from datetime import datetime, date
 import uuid
 from xml.etree import ElementTree
-from corehq.form_processor.interfaces import FormProcessorInterface
+from corehq.form_processor.interfaces.case import CaseInterface
+from corehq.form_processor.interfaces.processor import FormProcessorInterface
 from dimagi.utils.parsing import json_format_datetime
 from casexml.apps.case.xml import V1, NS_VERSION_MAP, V2
 from casexml.apps.case.const import DEFAULT_CASE_INDEX_IDENTIFIERS, CASE_INDEX_CHILD
@@ -314,7 +315,7 @@ class CaseFactory(object):
         return self.create_or_update_cases([case_structure], form_extras)
 
     def create_or_update_cases(self, case_structures, form_extras=None):
-        from corehq.form_processor.interfaces import FormProcessorInterface
+        from corehq.form_processor.interfaces.processor import FormProcessorInterface
 
         def _get_case_block(substructure):
             return self.get_case_block(substructure.case_id, index=substructure.index, **substructure.attrs)
@@ -333,6 +334,6 @@ class CaseFactory(object):
             form_extras,
         )
 
-        return FormProcessorInterface.get_cases(
+        return CaseInterface.get_cases(
             [id for structure in case_structures for id in structure.walk_ids()]
         )
