+++ corehq/ex-submodules/casexml/apps/case/tests/test_dbcache.py
@@ -5,7 +5,7 @@ from casexml.apps.case.mock import CaseBlock
 from casexml.apps.case.models import CommCareCase
 from casexml.apps.case.xform import CaseDbCache
 from casexml.apps.case.xml import V2
-from corehq.form_processor.interfaces import FormProcessorInterface
+from corehq.form_processor.interfaces.processor import FormProcessorInterface
 
 
 class CaseDbCacheTest(TestCase):
