+++ corehq/ex-submodules/casexml/apps/case/tests/test_out_of_order_processing.py
@@ -2,7 +2,8 @@ import os
 from django.test.utils import override_settings
 from django.test import TestCase
 from casexml.apps.case.tests.util import delete_all_cases
-from corehq.form_processor.interfaces import FormProcessorInterface
+from corehq.form_processor.interfaces.case import CaseInterface
+from corehq.form_processor.interfaces.processor import FormProcessorInterface
 
 
 @override_settings(CASEXML_FORCE_DOMAIN_CHECK=False)
@@ -19,7 +20,7 @@ class OutOfOrderCaseTest(TestCase):
 
             FormProcessorInterface.submit_form_locally(xml_data)
 
-        case = FormProcessorInterface.get_case('30bc51f6-3247-4966-b4ae-994f572e85fe')
+        case = CaseInterface.get_case('30bc51f6-3247-4966-b4ae-994f572e85fe')
         self.assertEqual('from the update form', case.pupdate)
         self.assertEqual('from the create form', case.pcreate)
         self.assertEqual('overridden by the update form', case.pboth)
