+++ corehq/ex-submodules/couchforms/tests/test_duplicates.py
@@ -1,7 +1,8 @@
 import os
 from django.test import TestCase
 
-from corehq.form_processor.interfaces import FormProcessorInterface
+from corehq.form_processor.interfaces.processor import FormProcessorInterface
+from corehq.form_processor.interfaces.xform import XFormInterface
 from corehq.form_processor.test_utils import FormProcessorTestUtils
 from corehq.form_processor.generic import GenericXFormInstance
 from couchforms.models import XFormInstance
@@ -49,7 +50,7 @@ class DuplicateFormTest(TestCase):
             doc_type='XFormInstance',
             domain='wrong-domain',
         )
-        xform = FormProcessorInterface.create_from_generic(generic_xform)
+        xform = XFormInterface.create_from_generic(generic_xform)
 
         instance = self._get_file()
         instance = instance.replace(self.ID, xform.id)
