+++ corehq/ex-submodules/couchforms/tests/test_xml.py
@@ -3,8 +3,9 @@
 
 import uuid
 import os
-from corehq.form_processor.interfaces import FormProcessorInterface
+from corehq.form_processor.interfaces.processor import FormProcessorInterface
 from django.test import TestCase
+from corehq.form_processor.interfaces.xform import XFormInterface
 
 
 class XMLElementTest(TestCase):
@@ -27,5 +28,5 @@ class XMLElementTest(TestCase):
             )
             xform = FormProcessorInterface.post_xform(xml_data)
             self.assertEqual(value, xform.form['test'])
-            elem = FormProcessorInterface.get_xml_element(xform)
+            elem = XFormInterface.get_xml_element(xform)
             self.assertEqual(value, elem.find('{http://commcarehq.org/couchforms-tests}test').text)
